+++ src/cuda/cuda_step.h
@@ -8,7 +8,8 @@
 
 using namespace std;
 
-void stepParticles(vector<double> &positionx, vector<double> &positiony , vector<double> &positionz,
-		   vector<double> &masses, vector<double> &velocityx, vector<double> &velocityy,
-		   vector<double> &velocityz, int h, int t);
+void stepParticles(vector<double> &positionx, vector<double> &positiony ,
+		   vector<double> &positionz, vector<double> &masses,
+		   vector<double> &velocityx, vector<double> &velocityy,
+		   vector<double> &velocityz, double h, double t);
 #endif
